date_18/04/2022

+++ src/controller/friendRequestController.ts
@@ -32,29 +32,44 @@ export let sendFriendRequest = async (
 
     if (!recieverData) {
       return next(new ApiError("no user found", 400));
-    } else {
-      const cheackFriendRequest = await conversation.findOne({
-        where: {
-          senderId: { [Op.or]: [req.id, numberId] },
-          recieverId: { [Op.or]: [req.id, numberId] },
-        },
-      });
+    }
 
-      if (cheackFriendRequest) {
-        return next(new ApiError("already send friend request", 400));
+    const cheackFriendRequest = await conversation.findOne({
+      where: {
+        senderId: { [Op.or]: [req.id, numberId] },
+        recieverId: { [Op.or]: [req.id, numberId] },
+      },
+    });
+
+    if (cheackFriendRequest) {
+      const senderId = cheackFriendRequest.dataValues.senderId;
+      const status = cheackFriendRequest.dataValues.state;
+      if (status === "blocked") {
+        return next(new ApiError("you can not send request", 400));
       }
-
+      if (status === "accepted") {
+        return next(new ApiError("you are already friend", 400));
+      }
+      if (status === "pending" && senderId === req.id) {
+        return next(new ApiError("you have alredy send request", 400));
+      }
+      if (status === "pending") {
+        return next(new ApiError("you have recieve request ", 400));
+      }
+    }
+    if (!cheackFriendRequest) {
       const createData = await conversation.create({
         senderId: req.id,
         recieverId: numberId,
       });
-
       return res.json({
         statusCode: 201,
         message: "friend request is send",
       });
     }
   } catch (e: any) {
+    console.log("eeeeeeeeee", e);
+
     return next(new ApiError(e.message, 400));
   }
 };
@@ -71,7 +86,7 @@ export let friendRequestAccept = async (
     if (Number.isNaN(parseInt(numberId)) === true) {
       return next(new ApiError("id type is string", 400));
     }
-    if (req.id == numberId) {
+    if (req.id == parseInt(numberId)) {
       return next(new ApiError("can not accept request", 404));
     }
 
@@ -82,38 +97,47 @@ export let friendRequestAccept = async (
     });
     if (!recieverData) {
       return next(new ApiError("no user found", 404));
-    } else {
-      const cheackFriendRequest = await conversation.findOne({
-        where: {
-          senderId: { [Op.or]: [req.id, numberId] },
-          recieverId: { [Op.or]: [req.id, numberId] },
-        },
-      });
-      if (!cheackFriendRequest) {
-        return next(new ApiError("no conversation found", 400));
+    }
+    const cheackFriendRequest = await conversation.findOne({
+      where: {
+        senderId: { [Op.or]: [req.id, parseInt(numberId)] },
+        recieverId: { [Op.or]: [req.id, parseInt(numberId)] },
+      },
+    });
+    if (!cheackFriendRequest) {
+      return next(new ApiError("no conversation found", 400));
+    }
+    if (cheackFriendRequest) {
+      const recieverId = cheackFriendRequest.dataValues.recieverId;
+      const match = req.id === recieverId;
+      const value = cheackFriendRequest.dataValues.state;
+      if (value === "pending" && match == false) {
+        return next(new ApiError("sender can not accept request", 400));
       }
-      if (cheackFriendRequest) {
-        const value = cheackFriendRequest.dataValues.isAccepted;
-        if (value == true) {
-          return next(new ApiError("already friend request accepted", 400));
-        }
 
-        const updateData = await conversation.update(
-          { isAccepted: true, isRejected: false },
+      if (value === "accepted" && match == true) {
+        return next(new ApiError("already accepted ", 400));
+      }
+
+      if (value === "pending" && match == true) {
+        console.log(req.id);
+        console.log(parseInt(numberId));
+
+        await conversation.update(
+          { state: "accepted" },
           {
             where: {
-              senderId: numberId,
+              // senderId: parseInt(numberId),
               recieverId: req.id,
             },
           }
         );
-        return res.json({
-          statusCode: 201,
-          message: "friend request is accepted",
-        });
+        return next(new ApiError("friend request is  accepted ", 200));
       }
     }
   } catch (e: any) {
+    console.log(e);
+
     return next(new ApiError(e.message, 400));
   }
 };
@@ -133,7 +157,7 @@ export let friendRequestReject = async (
 
     const recieverData: any = await users.findOne({
       where: {
-        id: numberId,
+        id: parseInt(numberId),
       },
     });
 
@@ -141,34 +165,33 @@ export let friendRequestReject = async (
       return next(new ApiError("can not reject request ", 404));
     }
 
-    if (!recieverData) {
-      return next(new ApiError("no user found", 404));
-    } else {
-      const cheackFriendRequest = await conversation.findOne({
-        where: {
-          senderId: { [Op.or]: [req.id, numberId] },
-          recieverId: { [Op.or]: [req.id, numberId] },
-        },
-      });
-      if (!cheackFriendRequest) {
-        return next(new ApiError("no conversation found ", 400));
+    const cheackFriendRequest = await conversation.findOne({
+      where: {
+        senderId: { [Op.or]: [req.id, parseInt(numberId)] },
+        recieverId: { [Op.or]: [req.id, parseInt(numberId)] },
+        state: "pending",
+      },
+    });
+
+    if (!cheackFriendRequest) {
+      return next(new ApiError("no conversation found", 400));
+    }
+
+    if (cheackFriendRequest) {
+      const recieverId = cheackFriendRequest.dataValues.recieverId;
+      const match = req.id === recieverId;
+      const value = cheackFriendRequest.dataValues.state;
+
+      if (value === "pending" && match == false) {
+        return next(new ApiError("sender can not reject request", 400));
       }
-      if (cheackFriendRequest) {
-        const value = cheackFriendRequest.dataValues.isRejected;
-        if (value === true) {
-          return next(new ApiError("already reject request", 400));
-        } else {
-          const DeleteData = await conversation.destroy({
-            where: {
-              senderId: numberId,
-              recieverId: req.id,
-            },
-          });
-          return res.json({
-            statusCode: 201,
-            message: "friend request is rejected",
-          });
-        }
+      if (value === "pending" && match == true) {
+        await conversation.destroy({
+          where: {
+            recieverId: req.id,
+          },
+        });
+        return next(new ApiError("friend request is  reject ", 200));
       }
     }
   } catch (e: any) {
@@ -187,7 +210,7 @@ export let seeFriendRequest = async (
     const userData = await conversation.findAll({
       where: {
         recieverId: id,
-        isAccepted: false,
+        state: "pending",
       },
       attributes: ["senderId"],
       include: [
@@ -208,8 +231,118 @@ export let seeFriendRequest = async (
       data: userData,
     });
   } catch (e: any) {
+    return next(new ApiError(e, 404));
+  }
+};
 
+export let seeSenderRequest = async (
+  req: any,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const id = req.id;
+
+    const userData = await conversation.findAll({
+      where: {
+        senderId: id,
+        isAccepted: false,
+      },
+      attributes: ["senderId"],
+      include: [
+        {
+          model: users,
+          attributes: ["email", "fullName", "id"],
+          as: "reciever",
+        },
+      ],
+    });
+
+    if (userData.length == 0) {
+      return next(new ApiError("no  friend  request is found", 404));
+    }
+
+    return res.json({
+      statusCode: 200,
+      data: userData,
+    });
+  } catch (e: any) {
     // if (e instanceof Error) = e.message
     return next(new ApiError(e, 404));
+  }
+};
+
+export let blockMessage = async (
+  req: any,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const id = req.params.id;
+    const numberId: any = id.replace(/[' "]+/g, "");
+    if (Number.isNaN(parseInt(numberId)) === true) {
+      return next(new ApiError("id type is string", 400));
+    }
+    const conversationData = await conversation.findOne({
+      where: {
+        [Op.or]: [
+          {
+            senderId: req.id,
+          },
+          {
+            recieverId: req.id,
+          },
+        ],
+      },
+    });
+    if (!conversationData) {
+      const connversationBlock = await conversation.create({
+        senderId: req.id,
+        recieverId: numberId,
+        state: "blocked",
+      });
+
+      return res.json({
+        statusCode: 200,
+        message: "user blocked successfully",
+      });
+    }
+    if (conversationData) {
+      const value = conversationData.dataValues.state;
+      if (value === "blocked") {
+        await conversationData.update(
+          { state: "accepted" },
+          {
+            where: {
+              senderId: req.id,
+              recieverId: numberId,
+            },
+          }
+        );
+        return res.json({
+          statusCode: 200,
+          message: "user un-blocked successfully",
+        });
+      } else {
+        await conversationData.update(
+          { state: "blocked" },
+          {
+            where: {
+              senderId: req.id,
+              recieverId: numberId,
+            },
+          }
+        );
+        return res.json({
+          statusCode: 200,
+          message: "user blocked successfully",
+        });
+      }
+    }
+  } catch (e: any) {
+    return res.json({
+      statusCode: 400,
+      message: e.message,
+    });
   }
 };
